commiteeeee

+++ data/database.py
@@ -1,9 +1,12 @@
 import json
 import os
 
-class Db:
-    def addClient(pessoa, tipo):
+class Banquinho:
+    def addClient(pessoa, senha):
         pessoaConvert = vars(pessoa)
+        senhaConvert = vars(senha)
+
+        '''carregando lista de clientes e atualizando arquivo''' 
         with open("clientes.json") as arquivo:
             clientsList = json.load(arquivo)
 
@@ -12,31 +15,59 @@ class Db:
         with open("clientes.json", "a") as arquivo:
             json.dump(clientsList, arquivo, indent = 4)
 
-        if tipo == "empresa":
-            nameArq = pessoa.cnpj
-        else:
-            nameArq = pessoa.nome + pessoa.cpf
+        '''criando arquivo individual'''
+        nomeArq = pessoa.cpf_cnpj
 
         with open(nameArq + ".json", "w") as arquivo:
             json.dump(pessoaConvert, arquivo, indent = 4)
 
-        
-    def rmClient(key, tipo):
+        '''arquivo de senhas'''
+        with open("senhas.json") as arquivo:
+            senhas = json.load(arquivo)
+
+        senhas.append(senhaConvert)
+
+
+    def rmClient(key):
         with open("clientes.json") as arquivo:
             clientsList = json.load(arquivo)
 
         for client in clientsList:
-            if client[tipo] == key:
-                clientsList.remove(client)
+            if client['cpf_cnpj'] == key:
+                if client['saldo'] != 0.00:
+                    return False
+                else:
+                    clientsList.remove(client)
 
-        with open("clientes.json", "w") as arquivo:
-            json.dump(clientsList, arquivo, indent = 4)
+                    with open("clientes.json", "w") as arquivo:
+                        json.dump(clientsList, arquivo, indent = 4)
         
-        os.remove(key + ".json")
+                    os.remove(key + ".json")
+
+                    break
 
     def opBancaria(operacao, key):
-        with open(key + ".json", "a") as arquivo:
-            json.dump(operacao, arquivo, indent = 4)
+        '''pegando o valor da operacao'''
+        if operacao.tipo == "Saque":
+            op = -operacao.valor
+        else:
+            op = operacao.valor
+        
+        '''carregando arquivo e atualizando saldo'''
+        with open(key + ".json") as arquivo:
+            cliente = json.load(arquivo)
+
+        saldo = cliente['saldo'] + op 
+        cliente['saldo'] = saldo
+        
+        '''adicionando operacao na lista de operacoes'''
+        cliente['operacoes'].append(operacao)
+
+        with open(key + ".json", "w") as arquivo:
+            json.dump(cliente, arquivo, indent = 4)
+
+        '''mexendo na lista geral'''
+
         
     def solCredito(dados):
         with open("pedidosDeCredito.json") as arquivo:
